requirements.txt
discord.py
sportsreference
feedparser
